+++ package/Runtime/ARCorePlaneExtension.cs
@@ -1,3 +1,4 @@
+using System;
 using UnityEngine.XR.ARExtensions;
 using UnityEngine.Experimental.XR;
 
@@ -12,11 +13,17 @@ namespace UnityEngine.XR.ARCore
         static void Register()
         {
             XRPlaneExtensions.RegisterGetTrackingStateHandler(k_SubsystemId, GetTrackingState);
+            XRPlaneExtensions.RegisterGetNativePtrHandler(k_SubsystemId, GetNativePtr);
         }
 
         static TrackingState GetTrackingState(XRPlaneSubsystem planeSubsystem, TrackableId planeId)
         {
             return Api.UnityARCore_getAnchorTrackingState(planeId);
+        }
+
+        static IntPtr GetNativePtr(XRPlaneSubsystem planeSubsystem, TrackableId planeId)
+        {
+            return Api.UnityARCore_getNativePlanePtr(planeId);
         }
 
         static readonly string k_SubsystemId = "ARCore-Plane";