+++ package/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-preview.22] - 2018-10-06
+### Improvements
+- Add linker validation when building with the IL2CPP scripting backend to avoid stripping the Unity.XR.ARCore assembly.
+- Add native pointer support for native AR objects
+
 ## [1.0.0-preview.21] - 2018-10-12
 ### Fixed
 - Fixed a bug which prevented the CameraImage API from working in 2018.3+